+++ diventi/accounts/models.py
@@ -4,22 +4,32 @@ from django.contrib.auth import models as auth_models
 from django.core.urlresolvers import reverse_lazy
 from django.utils.html import mark_safe
 
+from django.contrib.auth.models import BaseUserManager
+
 
 class DiventiUserManager(auth_models.BaseUserManager):
 
-    def create_user(self, email, password=None):
+    def _create_user(self, email, password, **extra_fields):
+        """
+        Creates and saves a User with the given email and password.
+        """
         if not email:
-            raise ValueError('Users must have an email address')
-        user = self.model(email=self.normalize_email(email))
+            raise ValueError('The Email must be set')
+        email = self.normalize_email(email)
+        user = self.model(email=email, **extra_fields)
         user.set_password(password)
-        user.save(using=self._db)
+        user.save()
         return user
 
-    def create_superuser(self, email, password):
-        user = self.create_user(email, password=password)
-        user.is_superuser = user.is_staff = True
-        user.save(using=self._db)
-        return user
+    def create_superuser(self, email, password, **extra_fields):
+        extra_fields.setdefault('is_staff', True)
+        extra_fields.setdefault('is_superuser', True)
+        extra_fields.setdefault('is_active', True)
+        if extra_fields.get('is_staff') is not True:
+            raise ValueError('Superuser must have is_staff=True.')
+        if extra_fields.get('is_superuser') is not True:
+            raise ValueError('Superuser must have is_superuser=True.')
+        return self._create_user(email, password, **extra_fields)
 
 
 class DiventiAvatarQuerySet(models.QuerySet):
@@ -52,17 +62,16 @@ class DiventiAvatar(models.Model):
         return u'{0}'.format(self.label)
 
 
-class DiventiUser(AbstractUser):
+class DiventiUser(AbstractUser):    
     email = models.EmailField(unique=True)
     avatar = models.ForeignKey(DiventiAvatar, null=True, related_name='diventiuser')
     profilepic = models.ImageField(blank=True, upload_to='profilepic/') # For the landing page, staff use only
     bio = models.TextField(blank=True)
-    role = models.CharField(blank=True, max_length=70)
-
-    objects = DiventiUserManager()
+    role = models.CharField(blank=True, max_length=70)    
 
     USERNAME_FIELD = 'email'
     REQUIRED_FIELDS = []
+    objects = DiventiUserManager()
 
     class Meta:
         verbose_name = 'User'
@@ -79,7 +88,7 @@ class DiventiUser(AbstractUser):
         return u'{0} {1}'.format(self.first_name, self.last_name)
 
     def clean(self):
-        """ Set email as username before saving a user """
+        # Set the first part of the email as username
         self.username = self.email
 
     def __str__(self):